+++ mm2src/coins/lightning/ln_platform.rs
@@ -8,7 +8,7 @@ use crate::utxo::utxo_common::UtxoTxBuilder;
 use crate::utxo::utxo_standard::UtxoStandardCoin;
 use crate::utxo::{FeePolicy, GenerateTxError, GenerateTxResult, GetConfirmedTxError, GetUtxoListOps,
                   UtxoTxGenerationOps};
-use crate::{CoinFutSpawner, MarketCoinOps, MmCoin};
+use crate::{CoinFutSpawner, MarketCoinOps, MmCoin, WaitForHTLCTxSpendArgs};
 use bitcoin::blockdata::block::BlockHeader;
 use bitcoin::blockdata::script::Script;
 use bitcoin::blockdata::transaction::Transaction;
@@ -18,6 +18,7 @@ use bitcoin_hashes::{sha256d, Hash};
 use chain::TransactionOutput;
 use common::executor::{abortable_queue::AbortableQueue, AbortableSystem, SpawnFuture, Timer};
 use common::log::{debug, error, info};
+use common::wait_until_sec;
 use futures::compat::Future01CompatExt;
 use futures::future::join_all;
 use keys::hash::H256;
@@ -408,7 +409,7 @@ impl Platform {
                 if is_spv_enabled {
                     client
                         // TODO: Should log the spv error if height > 0
-                        .validate_spv_proof(&transaction, (now_ms() / 1000) + TRY_SPV_PROOF_INTERVAL)
+                        .validate_spv_proof(&transaction, wait_until_sec(TRY_SPV_PROOF_INTERVAL))
                         .await
                         .map_err(GetConfirmedTxError::SPVError)
                 } else {
@@ -478,7 +479,7 @@ impl Platform {
                 if is_spv_enabled {
                     client
                         // TODO: Should log the spv error if height > 0
-                        .validate_spv_proof(&output.spending_tx, (now_ms() / 1000) + TRY_SPV_PROOF_INTERVAL)
+                        .validate_spv_proof(&output.spending_tx, wait_until_sec(TRY_SPV_PROOF_INTERVAL))
                         .await
                         .map_err(GetConfirmedTxError::SPVError)
                 } else {
@@ -582,14 +583,15 @@ impl Platform {
         let closing_tx = self
             .coin
             // TODO add fn with old wait_for_tx_spend name
-            .wait_for_htlc_tx_spend(
-                &funding_tx_bytes.into_vec(),
-                &[],
-                (now_ms() / 1000) + 3600,
+            .wait_for_htlc_tx_spend(WaitForHTLCTxSpendArgs {
+                tx_bytes: &funding_tx_bytes.into_vec(),
+                secret_hash: &[],
+                wait_until: wait_until_sec(3600),
                 from_block,
-                &None,
-                TAKER_PAYMENT_SPEND_SEARCH_INTERVAL,
-            )
+                swap_contract_address: &None,
+                check_every: TAKER_PAYMENT_SPEND_SEARCH_INTERVAL,
+                watcher_reward: false,
+            })
             .compat()
             .await
             .map_to_mm(|e| SaveChannelClosingError::WaitForFundingTxSpendError(e.get_plain_text_format()))?;