+++ mm2src/coins/lp_coins.rs
@@ -46,8 +46,8 @@ use base58::FromBase58Error;
 use common::custom_futures::timeout::TimeoutError;
 use common::executor::{abortable_queue::{AbortableQueue, WeakSpawner},
                        AbortSettings, AbortedError, SpawnAbortable, SpawnFuture};
-use common::log::LogOnError;
-use common::{calc_total_pages, now_ms, ten, HttpStatusCode};
+use common::log::{warn, LogOnError};
+use common::{calc_total_pages, now_sec, ten, HttpStatusCode};
 use crypto::{Bip32Error, CryptoCtx, CryptoCtxError, DerivationPath, GlobalHDAccountArc, HwRpcError, KeyPairPolicy,
              Secp256k1Secret, WithHwRpcError};
 use derive_more::Display;
@@ -76,6 +76,8 @@ use std::future::Future as Future03;
 use std::num::NonZeroUsize;
 use std::ops::{Add, Deref};
 use std::str::FromStr;
+use std::sync::atomic::AtomicBool;
+use std::sync::atomic::Ordering as AtomicOrdering;
 use std::sync::Arc;
 use std::time::Duration;
 use utxo_signer::with_key_pair::UtxoSignWithKeyPairError;
@@ -199,6 +201,8 @@ macro_rules! ok_or_continue_after_sleep {
 }
 
 pub mod coin_balance;
+pub mod lp_price;
+pub mod watcher_common;
 
 pub mod coin_errors;
 use coin_errors::{MyAddressError, ValidatePaymentError, ValidatePaymentFut};
@@ -319,7 +323,6 @@ pub const INVALID_PAYMENT_STATE_ERR_LOG: &str = "Invalid payment state";
 pub const INVALID_SWAP_ID_ERR_LOG: &str = "Invalid swap id";
 pub const INVALID_SCRIPT_ERR_LOG: &str = "Invalid script";
 pub const INVALID_REFUND_TX_ERR_LOG: &str = "Invalid refund transaction";
-pub const INSUFFICIENT_WATCHER_REWARD_ERR_LOG: &str = "Insufficient watcher reward";
 
 #[derive(Debug, Deserialize, Display, Serialize, SerializeErrorType)]
 #[serde(tag = "error_type", content = "error_data")]
@@ -586,7 +589,7 @@ pub struct WatcherValidateTakerFeeInput {
     pub lock_duration: u64,
 }
 
-#[derive(Clone, Debug)]
+#[derive(Clone)]
 pub struct WatcherValidatePaymentInput {
     pub payment_tx: Vec<u8>,
     pub taker_payment_refund_preimage: Vec<u8>,
@@ -594,9 +597,9 @@ pub struct WatcherValidatePaymentInput {
     pub taker_pub: Vec<u8>,
     pub maker_pub: Vec<u8>,
     pub secret_hash: Vec<u8>,
-    pub try_spv_proof_until: u64,
+    pub wait_until: u64,
     pub confirmations: u64,
-    pub min_watcher_reward: Option<u64>,
+    pub maker_coin: MmCoinEnum,
 }
 
 #[derive(Clone, Debug)]
@@ -611,7 +614,7 @@ pub struct ValidatePaymentInput {
     pub try_spv_proof_until: u64,
     pub confirmations: u64,
     pub unique_swap_data: Vec<u8>,
-    pub min_watcher_reward: Option<u64>,
+    pub watcher_reward: Option<WatcherReward>,
 }
 
 #[derive(Clone, Debug)]
@@ -645,6 +648,23 @@ pub struct SearchForSwapTxSpendInput<'a> {
     pub watcher_reward: bool,
 }
 
+#[derive(Copy, Clone, Debug)]
+pub enum RewardTarget {
+    None,
+    Contract,
+    PaymentSender,
+    PaymentSpender,
+    PaymentReceiver,
+}
+
+#[derive(Clone, Debug)]
+pub struct WatcherReward {
+    pub amount: BigDecimal,
+    pub is_exact_amount: bool,
+    pub reward_target: RewardTarget,
+    pub send_contract_reward_on_spend: bool,
+}
+
 #[derive(Clone, Debug)]
 pub struct SendPaymentArgs<'a> {
     pub time_lock_duration: u64,
@@ -658,7 +678,7 @@ pub struct SendPaymentArgs<'a> {
     pub swap_contract_address: &'a Option<BytesJson>,
     pub swap_unique_data: &'a [u8],
     pub payment_instructions: &'a Option<PaymentInstructions>,
-    pub watcher_reward: Option<u64>,
+    pub watcher_reward: Option<WatcherReward>,
     pub wait_for_confirmation_until: u64,
 }
 
@@ -725,15 +745,38 @@ pub struct EthValidateFeeArgs<'a> {
     pub uuid: &'a [u8],
 }
 
+#[derive(Clone, Debug)]
+pub struct WaitForHTLCTxSpendArgs<'a> {
+    pub tx_bytes: &'a [u8],
+    pub secret_hash: &'a [u8],
+    pub wait_until: u64,
+    pub from_block: u64,
+    pub swap_contract_address: &'a Option<BytesJson>,
+    pub check_every: f64,
+    pub watcher_reward: bool,
+}
+
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 pub enum PaymentInstructions {
     #[cfg(not(target_arch = "wasm32"))]
     Lightning(Invoice),
+    WatcherReward(BigDecimal),
+}
+
+#[derive(Clone, Debug, Default)]
+pub struct PaymentInstructionArgs<'a> {
+    pub secret_hash: &'a [u8],
+    pub amount: BigDecimal,
+    pub maker_lock_duration: u64,
+    pub expires_in: u64,
+    pub watcher_reward: bool,
+    pub wait_until: u64,
 }
 
 #[derive(Display)]
 pub enum PaymentInstructionsErr {
     LightningInvoiceErr(String),
+    WatcherRewardErr(String),
     InternalError(String),
 }
 
@@ -745,6 +788,7 @@ impl From<NumConversError> for PaymentInstructionsErr {
 pub enum ValidateInstructionsErr {
     ValidateLightningInvoiceErr(String),
     UnsupportedCoin(String),
+    DeserializationErr(String),
 }
 
 #[cfg(not(target_arch = "wasm32"))]
@@ -758,6 +802,13 @@ pub enum RefundError {
     DbError(String),
     Timeout(String),
     Internal(String),
+}
+
+#[derive(Debug, Display)]
+pub enum WatcherRewardError {
+    RPCError(String),
+    InvalidCoinType(String),
+    InternalError(String),
 }
 
 /// Swap operations (mostly based on the Hash/Time locked transactions implemented by coin wallets).
@@ -811,7 +862,7 @@ pub trait SwapOps {
     /// For example: there are no additional conditions for ETH, but for some UTXO coins we should wait for
     /// locktime < MTP
     fn can_refund_htlc(&self, locktime: u64) -> Box<dyn Future<Item = CanRefundHtlc, Error = String> + Send + '_> {
-        let now = now_ms() / 1000;
+        let now = now_sec();
         let result = if now > locktime {
             CanRefundHtlc::CanRefundNow
         } else {
@@ -843,33 +894,25 @@ pub trait SwapOps {
     /// Instructions from the taker on how the maker should send his payment.
     async fn maker_payment_instructions(
         &self,
-        secret_hash: &[u8],
-        amount: &BigDecimal,
-        maker_lock_duration: u64,
-        expires_in: u64,
+        args: PaymentInstructionArgs<'_>,
     ) -> Result<Option<Vec<u8>>, MmError<PaymentInstructionsErr>>;
 
     /// Instructions from the maker on how the taker should send his payment.
     async fn taker_payment_instructions(
         &self,
-        secret_hash: &[u8],
-        amount: &BigDecimal,
-        expires_in: u64,
+        args: PaymentInstructionArgs<'_>,
     ) -> Result<Option<Vec<u8>>, MmError<PaymentInstructionsErr>>;
 
     fn validate_maker_payment_instructions(
         &self,
         instructions: &[u8],
-        secret_hash: &[u8],
-        amount: BigDecimal,
-        maker_lock_duration: u64,
+        args: PaymentInstructionArgs<'_>,
     ) -> Result<PaymentInstructions, MmError<ValidateInstructionsErr>>;
 
     fn validate_taker_payment_instructions(
         &self,
         instructions: &[u8],
-        secret_hash: &[u8],
-        amount: BigDecimal,
+        args: PaymentInstructionArgs<'_>,
     ) -> Result<PaymentInstructions, MmError<ValidateInstructionsErr>>;
 
     fn is_supported_by_watchers(&self) -> bool { false }
@@ -931,6 +974,22 @@ pub trait WatcherOps {
         &self,
         input: WatcherSearchForSwapTxSpendInput<'_>,
     ) -> Result<Option<FoundSwapTxSpend>, String>;
+
+    async fn get_taker_watcher_reward(
+        &self,
+        other_coin: &MmCoinEnum,
+        coin_amount: Option<BigDecimal>,
+        other_coin_amount: Option<BigDecimal>,
+        reward_amount: Option<BigDecimal>,
+        wait_until: u64,
+    ) -> Result<WatcherReward, MmError<WatcherRewardError>>;
+
+    async fn get_maker_watcher_reward(
+        &self,
+        other_coin: &MmCoinEnum,
+        reward_amount: Option<BigDecimal>,
+        wait_until: u64,
+    ) -> Result<Option<WatcherReward>, MmError<WatcherRewardError>>;
 }
 
 /// Operations that coins have independently from the MarketMaker.
@@ -977,15 +1036,7 @@ pub trait MarketCoinOps {
 
     fn wait_for_confirmations(&self, input: ConfirmPaymentInput) -> Box<dyn Future<Item = (), Error = String> + Send>;
 
-    fn wait_for_htlc_tx_spend(
-        &self,
-        transaction: &[u8],
-        secret_hash: &[u8],
-        wait_until: u64,
-        from_block: u64,
-        swap_contract_address: &Option<BytesJson>,
-        check_every: f64,
-    ) -> TransactionFut;
+    fn wait_for_htlc_tx_spend(&self, args: WaitForHTLCTxSpendArgs<'_>) -> TransactionFut;
 
     fn tx_enum_from_bytes(&self, bytes: &[u8]) -> Result<TransactionEnum, MmError<TxMarshalingErr>>;
 
@@ -2307,7 +2358,7 @@ pub trait MmCoin:
     fn on_disabled(&self) -> Result<(), AbortedError>;
 
     /// For Handling the removal/deactivation of token on platform coin deactivation.
-    fn on_token_deactivated(&self, _ticker: &str);
+    fn on_token_deactivated(&self, ticker: &str);
 }
 
 /// The coin futures spawner. It's used to spawn futures that can be aborted immediately or after a timeout
@@ -2496,6 +2547,8 @@ impl MmCoinEnum {
     }
 
     pub fn is_eth(&self) -> bool { matches!(self, MmCoinEnum::EthCoin(_)) }
+
+    fn is_platform_coin(&self) -> bool { self.ticker() == self.platform_ticker() }
 }
 
 #[async_trait]
@@ -2503,10 +2556,51 @@ pub trait BalanceTradeFeeUpdatedHandler {
     async fn balance_updated(&self, coin: &MmCoinEnum, new_balance: &BigDecimal);
 }
 
+#[derive(Clone)]
+pub struct MmCoinStruct {
+    pub inner: MmCoinEnum,
+    is_available: Arc<AtomicBool>,
+}
+
+impl MmCoinStruct {
+    fn new(coin: MmCoinEnum) -> Self {
+        Self {
+            inner: coin,
+            is_available: AtomicBool::new(true).into(),
+        }
+    }
+
+    /// Gets the current state of the parent coin whether
+    /// it's available for the external requests or not.
+    ///
+    /// Always `true` for child tokens.
+    pub fn is_available(&self) -> bool {
+        !self.inner.is_platform_coin() // Tokens are always active or disabled
+            || self.is_available.load(AtomicOrdering::SeqCst)
+    }
+
+    /// Makes the coin disabled to the external requests.
+    /// Useful for executing `disable_coin` on parent coins
+    /// that have child tokens enabled.
+    ///
+    /// Ineffective for child tokens.
+    pub fn update_is_available(&self, to: bool) {
+        if !self.inner.is_platform_coin() {
+            warn!(
+                "`update_is_available` is ineffective for tokens. Current token: {}",
+                self.inner.ticker()
+            );
+            return;
+        }
+
+        self.is_available.store(to, AtomicOrdering::SeqCst);
+    }
+}
+
 pub struct CoinsContext {
     /// A map from a currency ticker symbol to the corresponding coin.
     /// Similar to `LP_coins`.
-    coins: AsyncMutex<HashMap<String, MmCoinEnum>>,
+    coins: AsyncMutex<HashMap<String, MmCoinStruct>>,
     balance_update_handlers: AsyncMutex<Vec<Box<dyn BalanceTradeFeeUpdatedHandler + Send + Sync>>>,
     account_balance_task_manager: AccountBalanceTaskManagerShared,
     create_account_manager: CreateAccountTaskManagerShared,
@@ -2553,6 +2647,7 @@ impl CoinsContext {
                 coin: coin.ticker().into(),
             });
         }
+
         let ticker = coin.ticker();
 
         let mut platform_coin_tokens = self.platform_coin_tokens.lock();
@@ -2561,7 +2656,8 @@ impl CoinsContext {
             platform.insert(ticker.to_owned());
         }
 
-        coins.insert(ticker.into(), coin);
+        coins.insert(ticker.into(), MmCoinStruct::new(coin));
+
         Ok(())
     }
 
@@ -2579,14 +2675,19 @@ impl CoinsContext {
         let mut coins = self.coins.lock().await;
         let mut platform_coin_tokens = self.platform_coin_tokens.lock();
 
-        if coins.contains_key(platform.ticker()) {
-            return MmError::err(PlatformIsAlreadyActivatedErr {
-                ticker: platform.ticker().into(),
-            });
-        }
+        let platform_ticker = platform.ticker().to_owned();
 
-        let platform_ticker = platform.ticker().to_string();
-        coins.insert(platform_ticker.clone(), platform);
+        if let Some(coin) = coins.get(&platform_ticker) {
+            if coin.is_available() {
+                return MmError::err(PlatformIsAlreadyActivatedErr {
+                    ticker: platform.ticker().into(),
+                });
+            }
+
+            coin.update_is_available(true);
+        } else {
+            coins.insert(platform_ticker.clone(), MmCoinStruct::new(platform));
+        }
 
         // Tokens can't be activated without platform coin so we can safely insert them without checking prior existence
         let mut token_tickers = Vec::with_capacity(tokens.len());
@@ -2596,8 +2697,12 @@ impl CoinsContext {
         // We try to activate ETH coin and USDT token via enable_eth_with_tokens
         for token in tokens {
             token_tickers.push(token.ticker().to_string());
-            coins.insert(token.ticker().into(), token);
+            coins
+                .entry(token.ticker().into())
+                .or_insert_with(|| MmCoinStruct::new(token));
         }
+
+        token_tickers.dedup();
 
         platform_coin_tokens
             .entry(platform_ticker)
@@ -2625,6 +2730,7 @@ impl CoinsContext {
                     if let Some(token) = coins_storage.remove(token) {
                         // Abort all token related futures on token deactivation
                         token
+                            .inner
                             .on_disabled()
                             .error_log_with_msg(&format!("Error aborting coin({ticker}) futures"));
                     }
@@ -2635,7 +2741,7 @@ impl CoinsContext {
                 tokens.remove(ticker);
             }
             if let Some(platform_coin) = coins_storage.get(platform_ticker) {
-                platform_coin.on_token_deactivated(ticker);
+                platform_coin.inner.on_token_deactivated(ticker);
             }
         };
 
@@ -3107,10 +3213,10 @@ pub async fn lp_register_coin(
         RawEntryMut::Occupied(_oe) => {
             return MmError::err(RegisterCoinError::CoinIsInitializedAlready { coin: ticker.clone() })
         },
-        RawEntryMut::Vacant(ve) => ve.insert(ticker.clone(), coin.clone()),
+        RawEntryMut::Vacant(ve) => ve.insert(ticker.clone(), MmCoinStruct::new(coin.clone())),
     };
 
-    if coin.ticker() == coin.platform_ticker() {
+    if coin.is_platform_coin() {
         let mut platform_coin_tokens = cctx.platform_coin_tokens.lock();
         platform_coin_tokens
             .entry(coin.ticker().to_string())
@@ -3132,6 +3238,21 @@ fn lp_spawn_tx_history(ctx: MmArc, coin: MmCoinEnum) -> Result<(), String> {
 pub async fn lp_coinfind(ctx: &MmArc, ticker: &str) -> Result<Option<MmCoinEnum>, String> {
     let cctx = try_s!(CoinsContext::from_ctx(ctx));
     let coins = cctx.coins.lock().await;
+
+    if let Some(coin) = coins.get(ticker) {
+        if coin.is_available() {
+            return Ok(Some(coin.inner.clone()));
+        }
+    };
+
+    Ok(None)
+}
+
+/// Returns coins even if they are on the passive mode
+pub async fn lp_coinfind_any(ctx: &MmArc, ticker: &str) -> Result<Option<MmCoinStruct>, String> {
+    let cctx = try_s!(CoinsContext::from_ctx(ctx));
+    let coins = cctx.coins.lock().await;
+
     Ok(coins.get(ticker).cloned())
 }
 
@@ -3425,7 +3546,7 @@ pub async fn get_enabled_coins(ctx: MmArc) -> Result<Response<Vec<u8>>, String> 
     let enabled_coins: Vec<_> = try_s!(coins
         .iter()
         .map(|(ticker, coin)| {
-            let address = try_s!(coin.my_address());
+            let address = try_s!(coin.inner.my_address());
             Ok(EnabledCoin {
                 ticker: ticker.clone(),
                 address,
@@ -3909,4 +4030,62 @@ fn coins_conf_check(ctx: &MmArc, coins_en: &Json, ticker: &str, req: Option<&Jso
         );
     }
     Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use common::block_on;
+    use mm2_test_helpers::for_tests::RICK;
+
+    #[test]
+    fn test_lp_coinfind() {
+        let ctx = mm2_core::mm_ctx::MmCtxBuilder::default().into_mm_arc();
+        let coins_ctx = CoinsContext::from_ctx(&ctx).unwrap();
+        let coin = MmCoinEnum::Test(TestCoin::new(RICK));
+
+        // Add test coin to coins context
+        common::block_on(coins_ctx.add_platform_with_tokens(coin.clone(), vec![])).unwrap();
+
+        // Try to find RICK from coins context that was added above
+        let _found = common::block_on(lp_coinfind(&ctx, RICK)).unwrap();
+
+        assert!(matches!(Some(coin), _found));
+
+        block_on(coins_ctx.coins.lock())
+            .get(RICK)
+            .unwrap()
+            .update_is_available(false);
+
+        // Try to find RICK from coins context after making it passive
+        let found = common::block_on(lp_coinfind(&ctx, RICK)).unwrap();
+
+        assert!(found.is_none());
+    }
+
+    #[test]
+    fn test_lp_coinfind_any() {
+        let ctx = mm2_core::mm_ctx::MmCtxBuilder::default().into_mm_arc();
+        let coins_ctx = CoinsContext::from_ctx(&ctx).unwrap();
+        let coin = MmCoinEnum::Test(TestCoin::new(RICK));
+
+        // Add test coin to coins context
+        common::block_on(coins_ctx.add_platform_with_tokens(coin.clone(), vec![])).unwrap();
+
+        // Try to find RICK from coins context that was added above
+        let _found = common::block_on(lp_coinfind_any(&ctx, RICK)).unwrap();
+
+        assert!(matches!(Some(coin.clone()), _found));
+
+        block_on(coins_ctx.coins.lock())
+            .get(RICK)
+            .unwrap()
+            .update_is_available(false);
+
+        // Try to find RICK from coins context after making it passive
+        let _found = common::block_on(lp_coinfind_any(&ctx, RICK)).unwrap();
+
+        assert!(matches!(Some(coin), _found));
+    }
 }